requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
python-dotenv
flake8
